requirements.txt
spotdl
ffmpeg-python
yt-dlp
python-dotenv